Create Linked List
A linked List (LL) is a linear Data Structure.

See the implementation of LL in JavaScript. You will be amazed to discover that browsers use DLL (doubly linked List) for the back and next page navigation of webpages, and even music players use DLL for the next and previous tracks.

In the end, you will see LL implementation is a nested Objects structure when using JavaScript.
Have fun with it.